fix: fix tsconfig
Paths in tsconfig are relative to the config fine in which they are declared.
So any path declared in `packages/tsconfig/base.json` is relative to that directory.
This applies in particular to `outDir` and `include`. These have to be redeclared
in each `tsconfig.json` file in each library.
See https://github.com/microsoft/TypeScript/issues/29172